requirements.txt
flask==2.3.2
flask-pymongo==2.3.0
pymongo==4.3.3
python-dotenv==1.0.0
werkzeug==2.3.4
marshmallow==3.19.0
PyJWT==2.7.0
flask-cors==4.0.0
firebase-admin